auth/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "i-app-auth-service"
version = "1.0.0"
description = "Authentication microservice for I App"
authors = [{name = "I App Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi==0.104.1",
    "gunicorn==21.2.0",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "psycopg2-binary==2.9.11",
    "asyncpg==0.29.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "cryptography==42.0.7",
    "PyJWT==2.9.0",
    "redis==5.0.1",
    "boto3==1.34.162",
    "prometheus-client==0.19.0",
    "structlog==24.1.0",
    "sentry-sdk[fastapi]==1.38.0",
    "python-dotenv==1.0.0",
    "email-validator==2.2.0",
    "httpx==0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
